+++ src/API/lastfm.js
@@ -4,40 +4,41 @@ const lastFmAPI = process.env.REACT_APP_LAST_FM_API;
 const lastFmCall =
   "https://ws.audioscrobbler.com/2.0/?format=json&api_key=" + lastFmAPI;
 
-async function getData(url, f) {
+
+async function getData(url) {
   try {
     const response = await axios.get(url);
-    const data = response.data;
-    let value;
-    f === undefined ? (value = data) : (value = f(data));
-    return value;
+    return response.data;
   } catch (error) {
     console.log(error);
   }
 }
 
-function getScrobbles(user) {
-  let pages = Math.ceil(user.playCount / 200);
-  let url = `${lastFmCall}&method=user.getrecenttracks&limit=200&user=${
-    this.state.user
-  }&page=`;
-  let scrobblesPromises = [];
-  let scrobbles = [];
-  for (let i = 1; i <= pages; i++) {
-    const returnRecentTracks = e => {
-      if (e.recenttracks.track[0]["@attr"]) {
-        e.recenttracks.track.shift();
-      }
-      return e.recenttracks.track;
-    };
-    scrobblesPromises.push(getData(`${url}${i}`, returnRecentTracks));
-  }
-  Promise.all(scrobblesPromises).then(values => {
-    scrobbles = scrobbles.concat(...values);
-    this.setState({
-      scrobbles: scrobbles
-    });
-  });
+// Exported functions used in UserPage
+
+/**
+ * Returns the number of scrobbles the user has
+ * @param {string} user Last.fm user
+ * @returns {Promise} Response to the method https://www.last.fm/api/show/user.getInfo
+ */
+function getUserPlayCount(user) {
+  let playCount = getData(lastFmCall + "&method=user.getinfo&user=" + user);
+  return playCount;
 }
 
-export default getData;
+/**
+ * Returns the specified page from the user's scrobble list. Each page has 200 entries.
+ * @param {string} user 
+ * @param {number} page Page of the
+ * @returns {Promise} Response to the method https://www.last.fm/api/show/user.getRecentTracks
+ */
+function getScrobbles(user, page) {
+  return getData(
+    `${lastFmCall}&method=user.getrecenttracks&limit=200&user=${user}&page=${page}`
+  );
+}
+
+module.exports = {
+  getScrobbles,
+  getUserPlayCount
+};