+++ CHANGELOG.md
@@ -10,6 +10,18 @@ the app functionality.
 
 ## [Unreleased]
 
+## [0.1.1] - 2019-06-29
+### Changed
+- Draw Heatmap incrementally, updating every time UserPage fetches another page of
+scrobbles
+- The Heatmap cells have an id based on the date they display. Previously, it was
+based on the days elapsed since the start of the displayed period.
+- Change the color palete of the Heatmap. 
+- Reduce the number of color tags from 8 to 4.
+
+### Fixed
+- Tooltip no longer goes offscreen but instead now flips to the left
+
 ## [0.1.0] - 2019-06-26
 ### Added
 - Create a React project
@@ -17,10 +29,12 @@ the app functionality.
 - Create a form to ask for the username to fetch
 - Get a list of all the scrobbles from the user in `user/:user`
 - Display a Heatmap of the last year (Today but one year ago, today)
+
 #### Meta
 - Create a basic GitHub repository
 - Add community files: templates, contributing, code of conduct.
 - Write a useful Readme
 
 [Unreleased]: https://github.com/davidomarf/lastfm/compare/v0.1.0..HEAD
+[0.1.1]: https://github.com/davidomarf/lastfm/compare/v0.1.1..HEAD
 [0.1.0]: https://github.com/davidomarf/lastfm/releases/tag/v0.1.0